+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.0 [unreleased]
 
 - Added basic support for new Google Charts loader
+- Added `configure` function
 - Dropped jQuery and Zepto dependencies for AJAX
 - Fixed legend colors on scatter chart for Chart.js